requirements.txt
Flask>=2.2.2
pycryptodome>=3.15.0
requests>=2.28.1
pyopenssl>=22.1.0